Reorganized Approve/Disapprove input validations
